# Refactor bilinear