… .pre-commit-config.yaml
… azure-pipelines.yml
… docs/source/grid.rst
… docs/source/swath.rst
… pyresample/bilinear/xarr.py
… pyresample/image.py
… pyresample/test/test_bilinear.py
… pyresample/test/test_image.py