This is a big refactoring for bilinear interpolation. Most of the bits are also re-used in the legacy Numpy version, but I didn't want to put too much time in pulling out all the parts of `pyresample.bilinear.get_sample_from_bil_info()`.

There will be some renaming and such that can be improved, but I've run out of steam for now :sweat_smile: Suggestions are wellcome!

 - [x] Closes #299  <!-- remove if there is no corresponding issue, which should only be the case for minor changes -->
 - [x] Tests added and updated <!-- for all bug fixes or enhancements -->
 - [x] Tests passed <!-- for all non-documentation changes -->
 - [x] Passes ``git diff origin/master **/*py | flake8 --diff`` <!-- remove if you did not edit any Python files  -->
 - [x] Fully documented <!-- remove if this change should not be visible to users, e.g., if it is an internal clean-up, or if this is part of a larger project that will be documented later -->

This refactoring also brings some performance improvements when using pre-computed resampling info. With the script below, using Satpy, I got the following timings:

*EDIT*: Timings updated for latest Satpy `master` (Oct-7 2020) with minimal other load on the laptop.

Pyresample `master` branch
- initial run (`generate=False`): ~3 m 28 s~ 3 m 14 s
- initial run (`generate=True`): ~4 m 42 s~ 4 m 20 s
- reusing the cached resampling data (`generate=False`): 18 s
- reusing the cached resampling data (`generate=True`): ~1 m 25 s~ 1 m 21 s

This PR:
- initial run (`generate=False`): ~2 m 50 s~ 1 m 36 s
- initial run (`generate=True`): ~3 m 30 s~ 2 m 14 s
- reusing the cached resampling data (`generate=False`): ~16 s~ 15 s
- reusing the cached resampling data (`generate=True`): ~48 s~ 45 s

```python
#!/usr/bin/env python

import os
os.environ['DASK_NUM_WORKERS'] = '2'
os.environ['OMP_NUM_THREADS'] = '1'

import glob
from satpy import Scene

def main():
    fnames = glob.glob('/home/lahtinep/data/satellite/geo/msg/*201611281100*__')

    glbl = Scene(reader='seviri_l1b_hrit', filenames=fnames)
    glbl.load(["natural_color", "fog", "overview",
               "hrv_clouds", "hrv_fog", "convection"],
              generate=True)
#              generate=False)
    lcl = glbl.resample('euro4', resampler="bilinear", cache_dir="/tmp", reduce_data=False)
    lcl.save_datasets(base_dir='/tmp')


if __name__ == "__main__":
    main()

```