@pnuu 2020-09-09
I'll fix the failing tests for `image.py` functions tomorrow.