@pnuu 2020-09-10
Ok, now the tests pass locally. I didn't bother fixing the styling in `pyresample.image`, which I now deprecated.