@pnuu 2020-09-10
I'm going to ignore the DeepCode errors/warnings. The access to private attributes are within the tests, and in deprecated functions (`resample_bilinear()`, `get_bil_info()` and `get_sample_from_bil_info()`) for which I don't want to change the names of the attributes.

The `resample_bilinear()` function should be replaced with `.resample()` convenience method in a later PR.