@pnuu 2020-09-14
Any other suggestions?

@djhoese made a PR for Satpy (https://github.com/pytroll/satpy/pull/1361) so that Windows testing would happen in Travis instead of Appveyor. Should we make similar change also to Pyresample, merge that, rebase/merge to this and see that also the Windows tests pass?