@djhoese 2020-09-14
How backwards compatible is this PR?

I would wait to turn off appveyor. Plus the Azure builds are failing for some other reason. Not sure why yet.