@pnuu 2020-09-14
This is fully backwards compatible, including both Numpy and XArray versions.

And again I mixed up Azure and Appveyor :roll_eyes: And yeah, didn't figure what's wrong with the **Azure** builds.