@pnuu 2020-09-14
Looking again at the logs again, it seems that Python 3.9 fails to install `pyproj` because `proj` executable isn't available:
```
created virtual environment CPython3.9.0.candidate.1-64 in 430ms
  creator CPython3Posix(dest=/tmp/tmp.uuC6OgfBeC/venv, clear=False, global=False)
  seeder FromAppData(download=False, pip=bundle, setuptools=bundle, wheel=bundle, via=copy, app_data_dir=/root/.local/share/virtualenv)
    added seed packages: pip==20.2.2, setuptools==49.6.0, wheel==0.35.1
  activators BashActivator,CShellActivator,FishActivator,PowerShellActivator,PythonActivator,XonshActivator
    + pip install /tmp/cibuildwheel/repaired_wheel/pyresample-1.16.0+96.g6e01eaa-cp39-cp39-manylinux2010_x86_64.whl
  ERROR: Command errored out with exit status 1:
   command: /tmp/tmp.uuC6OgfBeC/venv/bin/python /tmp/tmp.uuC6OgfBeC/venv/lib/python3.9/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /tmp/tmpe3vvkoco
       cwd: /tmp/pip-install-vgtgu41m/pyproj
  Complete output (1 lines):
  proj executable not found. Please set the PROJ_DIR variable.For more information see: https://pyproj4.github.io/pyproj/stable/installation.html
```