@djhoese 2020-09-14
I think I posted it on slack but did you ever turn off python 3.9? Check the skip variable at the top of the azure config and add an entry for the cp39 equivalent.