@pnuu 2020-09-15
With these two additional `da.compute()` calls the overall compute calls drop from 74 to 33 for the initial run of the above example script. At the same time, the processing time drops to around 2 m 35 s, 15 s less than earlier.