@pnuu 2020-10-07
I ran the timings again with the current Satpy `master` branch and minimal load on the laptop. The updated timings are in the description.